fix: avoid memory leak when generating thumbnails via external API. Additionally, file names are no longer based on ownerID plus the original upload name. We now generate a UUID-based filename and preserve only the extension using filepath.Ext, which avoids collisions and leaking user-specific information in stored media paths.

+++ internal/handler/media_handler.go
@@ -270,7 +270,7 @@ func (h *MediaHandler) UploadMedia(c *gin.Context) {
 	description := c.PostForm("description")
 
 	// Faz upload
-	media, err := h.service.UploadMedia(file, title, description, uint(userID))
+	media, err := h.service.UploadMedia(file, title, description, userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
